+++ frontend/app/src/store/reports/actions.ts
@@ -3,11 +3,17 @@ import i18n from '@/i18n';
 import { createTask, taskCompletion, TaskMeta } from '@/model/task';
 import { TaskType } from '@/model/task-type';
 import { api } from '@/services/rotkehlchen-api';
-import { Severity } from '@/store/notifications/consts';
-import { notify } from '@/store/notifications/utils';
-import { MUTATION_PROGRESS } from '@/store/reports/const';
+import {
+  emptyError,
+  MUTATION_PROGRESS,
+  MUTATION_REPORT_ERROR
+} from '@/store/reports/const';
 import { ReportState } from '@/store/reports/state';
-import { ReportProgress, TradeHistory } from '@/store/reports/types';
+import {
+  ReportError,
+  ReportProgress,
+  TradeHistory
+} from '@/store/reports/types';
 
 import { Message, RotkehlchenState } from '@/store/types';
 import { ProfitLossPeriod } from '@/typing/types';
@@ -19,6 +25,7 @@ export const actions: ActionTree<ReportState, RotkehlchenState> = {
       processingState: '',
       totalProgress: '0'
     } as ReportProgress);
+    commit(MUTATION_REPORT_ERROR, emptyError());
 
     const interval = setInterval(async () => {
       const progress = await api.history.getProgress();
@@ -58,6 +65,17 @@ export const actions: ActionTree<ReportState, RotkehlchenState> = {
       const { result } = await taskCompletion<TradeHistory, TaskMeta>(
         TaskType.TRADE_HISTORY
       );
+
+      if (!result || !result.overview || !result.allEvents) {
+        commit(MUTATION_REPORT_ERROR, {
+          error: '',
+          message: i18n
+            .t('actions.reports.generate.error.description', { error: '' })
+            .toString()
+        } as ReportError);
+        return;
+      }
+
       const { overview, allEvents } = result;
 
       const report = {
@@ -66,12 +84,10 @@ export const actions: ActionTree<ReportState, RotkehlchenState> = {
       };
       commit('set', report);
     } catch (e) {
-      notify(
-        e.message,
-        i18n.t('actions.reports.generate.error.title').toString(),
-        Severity.ERROR,
-        true
-      );
+      commit(MUTATION_REPORT_ERROR, {
+        error: e.message,
+        message: i18n.t('actions.reports.generate.error.description').toString()
+      } as ReportError);
     }
 
     clearInterval(interval);