+++ frontend/app/src/views/ProfitLossReport.vue
@@ -18,7 +18,15 @@
       </v-tooltip>
     </base-page-header>
     <generate v-show="!isRunning" @generate="generate($event)" />
-    <div v-if="loaded && !isRunning">
+    <error-screen
+      v-if="!isRunning && reportError.message"
+      class="mt-12"
+      :message="reportError.message"
+      :error="reportError.error"
+      :title="$t('profit_loss_report.error.title')"
+      :subtitle="$t('profit_loss_report.error.subtitle')"
+    />
+    <div v-if="loaded && !isRunning && !reportError.message">
       <v-row>
         <v-col>
           <i18n
@@ -78,6 +86,7 @@
 import { Component, Vue } from 'vue-property-decorator';
 import { mapGetters, mapState } from 'vuex';
 import BasePageHeader from '@/components/base/BasePageHeader.vue';
+import ErrorScreen from '@/components/error/ErrorScreen.vue';
 import ProgressScreen from '@/components/helper/ProgressScreen.vue';
 import AccountingSettingsDisplay from '@/components/profitloss/AccountingSettingsDisplay.vue';
 import Generate from '@/components/profitloss/Generate.vue';
@@ -85,12 +94,13 @@ import ProfitLossEvents from '@/components/profitloss/ProfitLossEvents.vue';
 import ProfitLossOverview from '@/components/profitloss/ProfitLossOverview.vue';
 import { Currency } from '@/model/currency';
 import { TaskType } from '@/model/task-type';
-import { ReportPeriod } from '@/store/reports/types';
+import { ReportError, ReportPeriod } from '@/store/reports/types';
 import { Message } from '@/store/types';
 import { AccountingSettings, ProfitLossPeriod } from '@/typing/types';
 
 @Component({
   components: {
+    ErrorScreen,
     ProfitLossOverview,
     ProfitLossEvents,
     AccountingSettingsDisplay,
@@ -101,7 +111,12 @@ import { AccountingSettings, ProfitLossPeriod } from '@/typing/types';
   computed: {
     ...mapGetters('tasks', ['isTaskRunning']),
     ...mapGetters('reports', ['progress', 'processingState']),
-    ...mapState('reports', ['loaded', 'accountingSettings', 'reportPeriod']),
+    ...mapState('reports', [
+      'loaded',
+      'accountingSettings',
+      'reportPeriod',
+      'reportError'
+    ]),
     ...mapGetters('session', ['currency'])
   }
 })
@@ -113,6 +128,7 @@ export default class ProfitLossReport extends Vue {
   processingState!: string;
   accountingSettings!: AccountingSettings;
   reportPeriod!: ReportPeriod;
+  reportError!: ReportError;
 
   get isRunning(): boolean {
     return this.isTaskRunning(TaskType.TRADE_HISTORY);
@@ -127,7 +143,7 @@ export default class ProfitLossReport extends Vue {
     try {
       if (this.$interop.isPackaged) {
         const directory = await this.$interop.openDirectory(
-          this.$tc('profit_loss_report.select_directory')
+          this.$t('profit_loss_report.select_directory').toString()
         );
         if (!directory) {
           return;
@@ -137,7 +153,7 @@ export default class ProfitLossReport extends Vue {
         const { success, message } = await this.$api.downloadCSV();
         if (!success) {
           this.showMessage(
-            message ?? this.$tc('profit_loss_report.download_failed')
+            message ?? this.$t('profit_loss_report.download_failed').toString()
           );
         }
       }
@@ -149,7 +165,7 @@ export default class ProfitLossReport extends Vue {
 
   private showMessage(description: string) {
     this.$store.commit('setMessage', {
-      title: this.$tc('profit_loss_report.csv_export_error'),
+      title: this.$t('profit_loss_report.csv_export_error').toString(),
       description: description
     } as Message);
   }